requirements.txt
alembic==1.16.4
amqp==5.3.1
aniso8601==10.0.1
apispec==6.8.2
asgiref==3.9.0
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backports-datetime-fromisoformat==2.0.3
bcrypt==3.2.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.6.15
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloud-init==24.4.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
drf-yasg==1.21.10
Flask==2.3.3
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
flask-restx==1.3.0
flask-smorest==0.42.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gunicorn==23.0.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.10
importlib-metadata==4.6.4
importlib_resources==6.5.2
incremental==21.3.0
inflection==0.5.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kombu==5.5.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2022.1
PyYAML==5.4.1
redis==6.2.0
requests==2.32.4
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
systemd-python==234
tomli==2.2.1
Twisted==22.1.0
typing_extensions==4.14.1
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
webargs==8.7.0
Werkzeug==2.3.8
zipp==1.0.0
zope.interface==5.4.0
